+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     * Change: Modify `get_forced_admin_color()` to not return an invalid admin color scheme
     * Change: Modify `set_forced_admin_color()` to not save an invalid admin color scheme
 * New: Add TODO.md and move existing TODO list from top of main plugin file into it (and add to it)
+* Change: Output currently forced admin color scheme beneath the color scheme picker for admins via a notice similar to an admin notice
 * Change: Tweak attribute spacing for `label` tag
 * Change: Use HTTPS for link to WP SVN repository in bin script for configuring unit tests
 * Change: Note compatibility through WP 5.4+